requirements.txt
odooku-odoo-addons>=12.0.0
odooku>=12.0.0